CQRS implementation and mediatr adjustments.

+++ src/Restaurants.Application/Mediator/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -6,14 +6,19 @@ using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Mediator.Restaurants.Commands.DeleteRestaurant;
 
-public class DeleteRestaurantCommandHandler(IRestaurantRepository restaurantRepository, ILogger<DeleteRestaurantCommandHandler> logger, IMapper mapper) : 
-    IRequestHandler<DeleteRestaurantCommand, RestaurantDto>
+public class DeleteRestaurantCommandHandler(IRestaurantRepository restaurantRepository, ILogger<DeleteRestaurantCommandHandler> logger) : 
+    IRequestHandler<DeleteRestaurantCommand, bool>
 {
-    public async Task<RestaurantDto> Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
+    public async Task<bool> Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
         logger.LogInformation($"Deleting restaurant with id {request.Id}...");
         var restaurant = await restaurantRepository.GetRestaurantByIdAsync( request.Id );
-        restaurantRepository.DeleteRestaurant(request.Id);
-        return mapper.Map<RestaurantDto>( restaurant );
+
+        if(restaurant is not null)
+        {
+            restaurantRepository.DeleteRestaurant(request.Id);
+            return true;
+        }
+        return false;
     }
 }
